Simplify the SimpleDB solutions.
Now you don't need to know what an iterator is - you can just strip suffixes and prefixes, and use 'contains' to find a single character.
