+++ exercise-book/src/simple-db.md
@@ -12,7 +12,6 @@ In this exercise, we will implement a toy protocol parser for a simple protocol 
 
 - create a safe protocol parser in Rust manually
 
-
 ## Prerequisites
 
 - basic pattern matching with `match`
@@ -25,7 +24,7 @@ In this exercise, we will implement a toy protocol parser for a simple protocol 
 
 1. Create a library project called `simple-db`.
 2. Implement appropriate data structures for `Command` and `Error`.
-3. Read the documentation for [`str`](https://doc.rust-lang.org/std/primitive.str.html), especially [`split_once()`](https://doc.rust-lang.org/std/primitive.str.html#method.split_once) and [`splitn()`](https://doc.rust-lang.org/std/primitive.str.html#method.splitn). Pay attention to their return type. Use the result value of `split_once()` and `splitn()` to guide your logic. The Step-by-Step-Solution contains a proposal.
+3. Read the documentation for [`str`](https://doc.rust-lang.org/std/primitive.str.html), especially [`strip_prefix()`](https://doc.rust-lang.org/std/primitive.str.html#method.strip_prefix) and [`strip_suffix()`](https://doc.rust-lang.org/std/primitive.str.html#method.strip_suffix). Pay attention to their return type.
 4. Implement the following function so that it implements the protocol specifications to parse the messages. Use the provided tests to help you with the case handling.
 
 ```rust, ignore
@@ -34,7 +33,7 @@ pub fn parse(input: &str) -> Result<Command, Error> {
 }
 ```
 
-The Step-by-Step-Solution contains steps 4a-e that explain a possible way to handle the cases in detail.
+The Step-by-Step-Solution contains steps 4a-c that explain a possible way to handle the cases in detail.
 
 ### Optional Tasks:
 
@@ -58,7 +57,7 @@ With the additional properties:
 
 2. A missing newline at the end of the command is an error.
 
-3. Data after the first newline is an error.
+3. A newline other than at the end of the command is an error.
 
 4. Empty payloads are allowed. In this case, the command is
     `PUBLISH \n`.
@@ -66,9 +65,9 @@ With the additional properties:
 Violations against the form of the messages and the properties are
 handled with the following error codes:
 
-- `TrailingData` (bytes found after newline)
+- `UnexpectedNewline` (a newline not at the end of the line)
 
-- `IncompleteMessage` (no newline)
+- `IncompleteMessage` (no newline at the end)
 
 - `EmptyMessage` (empty string instead of a command)
 
@@ -84,102 +83,5 @@ handled with the following error codes:
 Below are the tests your protocol parser needs to pass. You can copy them to the bottom of your `lib.rs`.
 
 ```rust, ignore
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    // Tests placement of \n
-    #[test]
-    fn test_missing_nl() {
-        let line = "RETRIEVE";
-        let result: Result<Command, Error> = parse(line);
-        let expected = Err(Error::IncompleteMessage);
-        assert_eq!(result, expected);
-    }
-    #[test]
-    fn test_trailing_data() {
-        let line = "PUBLISH The message\n is wrong \n";
-        let result: Result<Command, Error> = parse(line);
-        let expected = Err(Error::TrailingData);
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_empty_string() {
-        let line = "";
-        let result = parse(line);
-        let expected = Err(Error::IncompleteMessage);
-        assert_eq!(result, expected);
-    }
-
-    // Tests for empty messages and unknown commands
-
-    #[test]
-    fn test_only_nl() {
-        let line = "\n";
-        let result: Result<Command, Error> = parse(line);
-        let expected = Err(Error::EmptyMessage);
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_unknown_command() {
-        let line = "SERVE \n";
-        let result: Result<Command, Error> = parse(line);
-        let expected = Err(Error::UnknownCommand);
-        assert_eq!(result, expected);
-    }
-
-    // Tests correct formatting of RETRIEVE command
-
-    #[test]
-    fn test_retrieve_w_whitespace() {
-        let line = "RETRIEVE \n";
-        let result: Result<Command, Error> = parse(line);
-        let expected = Err(Error::UnexpectedPayload);
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_retrieve_payload() {
-        let line = "RETRIEVE this has a payload\n";
-        let result: Result<Command, Error> = parse(line);
-        let expected = Err(Error::UnexpectedPayload);
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_retrieve() {
-        let line = "RETRIEVE\n";
-        let result: Result<Command, Error> = parse(line);
-        let expected = Ok(Command::Retrieve);
-        assert_eq!(result, expected);
-    }
-
-    // Tests correct formatting of PUBLISH command
-
-    #[test]
-    fn test_publish() {
-        let line = "PUBLISH TestMessage\n";
-        let result: Result<Command, Error> = parse(line);
-        let expected = Ok(Command::Publish("TestMessage".into()));
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_empty_publish() {
-        let line = "PUBLISH \n";
-        let result: Result<Command, Error> = parse(line);
-        let expected = Ok(Command::Publish("".into()));
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_missing_payload() {
-        let line = "PUBLISH\n";
-        let result: Result<Command, Error> = parse(line);
-        let expected = Err(Error::MissingPayload);
-        assert_eq!(result, expected);
-    }
-}
+{{#include ../../exercise-solutions/simple-db/step4c/src/lib.rs:41:138}}
 ```
