+++ exercise-solutions/connected-mailbox/simple-db/src/lib.rs
@@ -16,8 +16,8 @@ pub enum Command {
 /// The ways in which this module can fail
 #[derive(Eq, PartialEq, Debug, thiserror::Error)]
 pub enum Error {
-    #[error("Unexpected data after the final newline")]
-    TrailingData,
+    #[error("Found a newline that wasn't at the end")]
+    UnexpectedNewline,
     #[error("Not enough data in command (missing newline?)")]
     IncompleteMessage,
     #[error("Input was empty")]
@@ -33,33 +33,26 @@ pub enum Error {
 /// Parse a command-string into `Command`, or produce an `Error` if it could not
 /// be parsed.
 pub fn parse(input: &str) -> Result<Command, Error> {
-    let message = match input.split_once('\n') {
-        Some((message, "")) => message,
-        Some(_) => return Err(Error::TrailingData),
-        None => return Err(Error::IncompleteMessage),
+    let Some(message) = input.strip_suffix('\n') else {
+        return Err(Error::IncompleteMessage);
     };
 
-    let mut substrings = message.splitn(2, ' ');
+    if message.contains('\n') {
+        return Err(Error::UnexpectedNewline);
+    }
 
-    // Note: `splitn` *always* returns at least one value
-    let command = substrings.next().unwrap();
-    match command {
-        "RETRIEVE" => {
-            if substrings.next().is_none() {
-                Ok(Command::Retrieve)
-            } else {
-                Err(Error::UnexpectedPayload)
-            }
-        }
-        "PUBLISH" => {
-            if let Some(payload) = substrings.next() {
-                Ok(Command::Publish(String::from(payload)))
-            } else {
-                Err(Error::MissingPayload)
-            }
-        }
-        "" => Err(Error::EmptyMessage),
-        _ => Err(Error::UnknownCommand),
+    if let Some(payload) = message.strip_prefix("PUBLISH ") {
+        Ok(Command::Publish(String::from(payload)))
+    } else if message == "PUBLISH" {
+        Err(Error::MissingPayload)
+    } else if message == "RETRIEVE" {
+        Ok(Command::Retrieve)
+    } else if let Some(_payload) = message.strip_prefix("RETRIEVE ") {
+        Err(Error::UnexpectedPayload)
+    } else if message == "" {
+        Err(Error::EmptyMessage)
+    } else {
+        Err(Error::UnknownCommand)
     }
 }
 
@@ -79,14 +72,14 @@ mod tests {
     fn test_trailing_data() {
         let line = "PUBLISH The message\n is wrong \n";
         let result: Result<Command, Error> = parse(line);
-        let expected = Err(Error::TrailingData);
+        let expected = Err(Error::UnexpectedNewline);
         assert_eq!(result, expected);
     }
 
     #[test]
     fn test_empty_string() {
         let line = "";
-        let result = parse(line);
+        let result: Result<Command, Error> = parse(line);
         let expected = Err(Error::IncompleteMessage);
         assert_eq!(result, expected);
     }
