+++ exercise-solutions/simple-db/step4b/src/lib.rs
@@ -2,12 +2,11 @@
 pub enum Command {
     Publish(String),
     Retrieve,
-    Command,
 }
 
 #[derive(Eq, PartialEq, Debug)]
 pub enum Error {
-    TrailingData,
+    UnexpectedNewline,
     IncompleteMessage,
     EmptyMessage,
     UnknownCommand,
@@ -16,17 +15,23 @@ pub enum Error {
 }
 
 pub fn parse(input: &str) -> Result<Command, Error> {
-    let message = match input.split_once('\n') {
-        Some((message, "")) => message,
-        Some(_) => return Err(Error::TrailingData),
-        None => return Err(Error::IncompleteMessage),
+    let Some(message) = input.strip_suffix('\n') else {
+        return Err(Error::IncompleteMessage);
     };
 
-    let mut substrings = message.splitn(2, ' ');
+    if message.contains('\n') {
+        return Err(Error::UnexpectedNewline);
+    }
 
-    let _command = substrings.next().unwrap();
-
-    Ok(Command::Command)
+    if message == "RETRIEVE" {
+        Ok(Command::Retrieve)
+    } else if let Some(_payload) = message.strip_prefix("RETRIEVE ") {
+        Err(Error::UnexpectedPayload)
+    } else if message == "" {
+        Err(Error::EmptyMessage)
+    } else {
+        Err(Error::UnknownCommand)
+    }
 }
 
 #[cfg(test)]
@@ -45,7 +50,59 @@ mod tests {
     fn test_trailing_data() {
         let line = "PUBLISH The message\n is wrong \n";
         let result: Result<Command, Error> = parse(line);
-        let expected = Err(Error::TrailingData);
+        let expected = Err(Error::UnexpectedNewline);
+        assert_eq!(result, expected);
+    }
+
+    #[test]
+    fn test_empty_string() {
+        let line = "";
+        let result: Result<Command, Error> = parse(line);
+        let expected = Err(Error::IncompleteMessage);
+        assert_eq!(result, expected);
+    }
+
+    // Tests for empty messages and unknown commands
+
+    #[test]
+    fn test_only_nl() {
+        let line = "\n";
+        let result: Result<Command, Error> = parse(line);
+        let expected = Err(Error::EmptyMessage);
+        assert_eq!(result, expected);
+    }
+
+    #[test]
+    fn test_unknown_command() {
+        let line = "SERVE \n";
+        let result: Result<Command, Error> = parse(line);
+        let expected = Err(Error::UnknownCommand);
+        assert_eq!(result, expected);
+    }
+
+    // Tests correct formatting of RETRIEVE command
+
+    #[test]
+    fn test_retrieve_w_whitespace() {
+        let line = "RETRIEVE \n";
+        let result: Result<Command, Error> = parse(line);
+        let expected = Err(Error::UnexpectedPayload);
+        assert_eq!(result, expected);
+    }
+
+    #[test]
+    fn test_retrieve_payload() {
+        let line = "RETRIEVE this has a payload\n";
+        let result: Result<Command, Error> = parse(line);
+        let expected = Err(Error::UnexpectedPayload);
+        assert_eq!(result, expected);
+    }
+
+    #[test]
+    fn test_retrieve() {
+        let line = "RETRIEVE\n";
+        let result: Result<Command, Error> = parse(line);
+        let expected = Ok(Command::Retrieve);
         assert_eq!(result, expected);
     }
 }
